change system configuration
